optionをセットする関数をsimul_search.goに統合
